requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
pymongo==3.12.1
djongo==1.3.6
uwsgi>=2.0.19.1